+++ examples/storybook/src/shared/utils/fileUploadUtil.ts
@@ -4,7 +4,9 @@ import type {
   UpdateEntityFunc,
   VideoComponentData,
 } from 'wix-rich-content-common';
+import type { AudioData as AudioComponentData } from 'ricos-schema';
 import { testImages, testWixVideos } from './mock';
+import { mockAudioData } from './mockAudioData';
 import { FILE_URLS } from './fileMockURL';
 
 export const mockImageNativeUploadFunc = (
@@ -111,6 +113,20 @@ export const mockCustomVideoUploadFunc = (
   }, 500);
 };
 
+export const mockCustomAudioUploadFunc = (
+  updateEntity: UpdateEntityFunc<AudioComponentData>,
+  removeEntity
+) => {
+  // eslint-disable-next-line no-console
+  console.log('consumer wants to upload custom audio');
+  const audioToUpload = getAudioToUpload('f0f74f_35a1cdce4973490eac49e74c3244364d');
+  setTimeout(() => {
+    updateEntity({ data: audioToUpload });
+    // eslint-disable-next-line no-console
+    console.log('consumer uploaded ', audioToUpload);
+  }, 500);
+};
+
 export const mockVideoNativeUploadFunc = (
   file: File,
   updateEntity: UpdateEntityFunc<VideoComponentData>,
@@ -125,6 +141,23 @@ export const mockVideoNativeUploadFunc = (
     updateEntity({ data: videoToUpload });
     // eslint-disable-next-line no-console
     console.log('consumer uploaded ', videoToUpload);
+  }, 5000);
+};
+
+export const mockAudioNativeUploadFunc = (
+  file: File,
+  updateEntity: UpdateEntityFunc<AudioComponentData>,
+  removeEntity
+) => {
+  // eslint-disable-next-line no-console
+  console.log('consumer wants to upload custom audio', file);
+  const mockAudioIndex = Math.floor(Math.random() * mockAudioData.length);
+  const testAudio = mockAudioData[mockAudioIndex];
+  const audioToUpload = getAudioToUpload(testAudio.url);
+  setTimeout(() => {
+    updateEntity({ data: audioToUpload });
+    // eslint-disable-next-line no-console
+    console.log('consumer uploaded ', audioToUpload);
   }, 5000);
 };
 
@@ -143,6 +176,20 @@ export const getVideoToUpload = (url: string, thumbnailUrl: string) => {
   // You can provide either absolute or relative URL.
   // If relative URL is provided, a function 'getVideoUrl' will be invoked to form a full URL.
   return videoWithRelativeUrl;
+};
+
+export const getAudioToUpload = (url: string) => {
+  const audioWithAbsoluteUrl = {
+    url: 'https://static.wixstatic.com/mp3/f0f74f_35a1cdce4973490eac49e74c3244364d.mp3',
+  };
+  const audioWithRelativeUrl = {
+    audio: { src: { id: `mp3/${url}.mp3` } },
+    name: 'Dear Fear',
+    authorName: 'KOTA The Friend',
+  };
+  // You can provide either absolute or relative URL.
+  // If relative URL is provided, a function 'getVideoUrl' will be invoked to form a full URL.
+  return audioWithRelativeUrl;
 };
 
 //////////////////////////////////////////FOR TESTS//////////////////////////////////////////