+++ examples/storybook/stories/Components/EditorWrapper.tsx
@@ -25,7 +25,7 @@ import { pluginMap, createMapPlugin } from 'wix-rich-content-plugin-map';
 import { pluginMentions } from 'wix-rich-content-plugin-mentions';
 import { pluginUndoRedo } from 'wix-rich-content-plugin-undo-redo';
 import { pluginVideo, videoButtonsTypes } from 'wix-rich-content-plugin-video';
-import { pluginAudio } from 'wix-rich-content-plugin-audio';
+import { pluginAudio, audioButtonsTypes } from 'wix-rich-content-plugin-audio';
 import { pluginPoll } from 'wix-rich-content-plugin-social-polls';
 import { pluginLinkPreview, LinkPreviewProviders } from 'wix-rich-content-plugin-link-preview';
 import {
@@ -224,6 +224,11 @@ const plugins = [
   }),
   pluginAudio({
     getAudioUrl: src => `https://static.wixstatic.com/${src.id}`,
+    exposeButtons: [
+      audioButtonsTypes.audio,
+      audioButtonsTypes.soundCloud,
+      audioButtonsTypes.spotify,
+    ],
   }),
   pluginLinkPreview(configs.linkPreview),
   pluginPoll(),