+++ packages/plugin-audio/web/src/index.ts
@@ -1,4 +1,4 @@
 export { createAudioPlugin } from './createAudioPlugin';
 export { default as audioViewer } from './audio-viewer';
-export { AUDIO_TYPE } from './types';
+export { AUDIO_TYPE, audioButtonsTypes } from './types';
 export { pluginAudio } from './editor';