+++ packages/plugin-audio/web/src/types.ts
@@ -1,6 +1,8 @@
 import type { EditorPluginConfig, ViewerPluginConfig } from 'wix-rich-content-common';
 export { AUDIO_TYPE } from 'ricos-content';
 
+export const audioButtonsTypes = { audio: 'audio', soundCloud: 'soundCloud', spotify: 'spotify' };
+
 export interface AudioPluginEditorConfig extends EditorPluginConfig {
   [key: string]: any; // eslint-disable-line @typescript-eslint/no-explicit-any
 }